version PFE_S32G_A53_LNX_RTM_1.1.0 RC1
Highlights:
    1) G3 Errors
    2) Health monitor
    3) Bugfixes

pfe.git: 1916c178aed42f8109de36393c4259186b19d5fc

+++ sw/pfe_hif_drv/public/pfe_hif_drv.h
@@ -430,7 +430,7 @@ static inline bool_t pfe_hif_pkt_icmp_csum_valid(const pfe_hif_pkt_t *pkt)
 	}
 #endif /* PFE_CFG_NULL_ARG_CHECK */
 
-	return !!(pkt->flags.specific.rx_flags & HIF_RX_ICMP_CSUM);
+	return !!((uint16_t)pkt->flags.specific.rx_flags & (uint16_t)HIF_RX_ICMP_CSUM);
 }
 /**
  * @brief		Get pointer to data buffer
